% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{AgingSubModel}
\alias{AgingSubModel}
\title{AgingSubModel}
\usage{
AgingSubModel(agent_states, step_data, step)
}
\arguments{
\item{agent_states}{agent_states object}

\item{step_data}{step_data object}

\item{step}{step object}
}
\value{
agent_states
}
\description{
AgingSubModel
}
