% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{AngleToPoint}
\alias{AngleToPoint}
\title{AngleToPoint}
\usage{
CenterXYInCell(x, y, xmin, ymin, cellsize)
}
\arguments{
\item{origin_x}{starting location x value}

\item{origin_y}{starting location y value}

\item{target_x}{finishing location x value}

\item{target_y}{finishing location y value}
}
\value{
numeric
}
\description{
Calculates the angle (based on a unit circle, where due East is 0 and due
West is pi) between an origin (x,y) and a target (x,y).
}
