% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{CalculateAngleToPoint}
\alias{CalculateAngleToPoint}
\title{CalculateAngleToPoint}
\usage{
CalculateAngleToPoint(origin_long, origin_lat, target_long, target_lat)
}
\arguments{
\item{origin_long}{column name of origin point longitude}

\item{origin_lat}{column name of origin point latitude}

\item{target_long}{column name of target point longitude}

\item{target_lat}{column name of target point latitude}
}
\value{
vector of angles
}
\description{
Calculates absolute angle in radians [0, 2*pi) of origin location to target
point
}
\details{
Coordinates must be in identically-scaled units (e.g. UTM meters).
If location data is in lat/long, project coordinates to UTM with 'rgdal'
package before running this function.
}
