% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{CalculateCurrentAge}
\alias{CalculateCurrentAge}
\title{CalculateCurrentAge}
\usage{
CalculateCurrentAge(datetime, birthdate)
}
\arguments{
\item{datetime}{datetime of step}

\item{birth_date}{birthdate of the agent}
}
\value{
current_age in days
}
\description{
Calculates current age of agent at a given step
}
