% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{CalculateReportAge}
\alias{CalculateReportAge}
\title{CalculateReportAge}
\usage{
CalculateReportAge(agent_states, step_data)
}
\arguments{
\item{agent_states}{= agent_states}

\item{step_data}{= step data dataframe}
}
\value{
report_date in 'report_age_period' time Period
}
\description{
Creates and updates population interval dataframe
}
\examples{

}
