% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{CenterXYInCell}
\alias{CenterXYInCell}
\title{CenterXYInCell}
\usage{
CenterXYInCell(x, y, xmin, ymin, cellsize)
}
\arguments{
\item{x}{x value}

\item{y}{y value}

\item{xmin}{minimum value of x that establishes the grid arrangement}

\item{ymin}{minimum value of y that establishes the grid arrangement}

\item{cellsize}{cell size in units of the x and y values}
}
\value{
vector of centered x and y (i.e., (x,y))
}
\description{
Centers x and y values into the center of a raster cell based on the xmin,
ymin, and cellsize parameters
}
