% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{CompileAllAgentsStepData}
\alias{CompileAllAgentsStepData}
\title{CompileAllAgentsStepData}
\usage{
CompileAllAgentsStepData(sim)
}
\arguments{
\item{sim}{a 'sim' list object}
}
\value{
a dataframe of all the step_data objects
}
\description{
Creates a compiled dataframe of all the agent$step_data objects within
sim$agents$all
}
\examples{

}
