% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{ConvertAngle}
\alias{ConvertAngle}
\title{ConvertAngle}
\usage{
ConvertAngle(x)
}
\arguments{
\item{x}{radian value}
}
\value{
vector of one value
}
\description{
Converts a radian angle that is outside of the Unit Circle range (i.e.,
  [0, 2pi]) to the equivalent value that is within the Unit Circle range.
}
\details{
Used so that the angles never go outside of the Unit Circle range
  when angles are being added or subtracted from each other over and over
  during a simulation.
}
