% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{ConvertStepDataCoordinates}
\alias{ConvertStepDataCoordinates}
\title{ConvertStepDataCoordinates}
\usage{
ConvertStepDataCoordinates(df)
}
\arguments{
\item{df}{input dataframe with "x", "y" columns}

\item{crs}{string of projection for "x", "y" columns, default is for
Maine BAEA GPS data (UTM Zone 19N).}
}
\value{
A dataframe with added "long", "lat" columns
}
\description{
Converts "x", "y" columns to coordinates to lat/long WGS84 (for Google)
}
