% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{CoordinatesFromAngleLength}
\alias{CoordinatesFromAngleLength}
\title{CoordinatesFromAngleLength}
\usage{
CoordinatesFromAngleLength(xy, angle, length)
}
\arguments{
\item{xy}{numeric, vector of two coordinates c(x,y)}

\item{angle}{numeric, radians [0, 2*pi) based on unit circle orientation.
Therefore, "due east" is 0, "due west" is pi.}

\item{length}{numeric, same units as x,y coordinates}
}
\value{
vector of two coordinates c(x,y)
}
\description{
Calculates a new x,y location based on an absolute angle in radians [0, 2*pi)
  and distance from origin x,y
}
\details{
Coordinates must be in identically-scaled units (e.g. UTM meters).
If location data is in lat & long, project coordinates to UTM with 'rgdal'
package before running this function.
}
