% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{CreateBase}
\alias{CreateBase}
\title{CreateBase}
\usage{
CreateAgentsInputClass(n, class, sex_ratio, age_min, age_max)
}
\arguments{
\item{x_min}{raster x minimum}

\item{y_min}{raster y minimum}

\item{x_max}{raster x maximum}

\item{y_max}{raster y maximum}

\item{cell_size}{integer, cell size (meters)}
}
\value{
RasterLayer
}
\description{
Creates a 'base' RasterLayer
}
