% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{CreateBirthDate}
\alias{CreateBirthDate}
\title{CreateBirthDate}
\usage{
CreateBirthDate(input)
}
\arguments{
\item{input}{a data frame containing the agents used to start the model.
Must containing a column specifying "age" as a numeric value.}
}
\value{
The input data frame with a column "birth_date" containing each
agent's birth_date as a POSIXct object.
}
\description{
Creates a column in the input data frame specifying each agent's birth date
baed on its specified age
}
