% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{CreateConNestProb}
\alias{CreateConNestProb}
\title{CreateConNestProb}
\usage{
CreateConNestProb(con_nest_raster, gamma_shape, gamma_rate, x, y, max_r,
  cellsize, base)
}
\arguments{
\item{con_nest_raster}{Raster,}

\item{gamma_shape}{Numeric}

\item{gamma_rate}{Numeric}

\item{x}{Numeric,}

\item{y}{Numeric,}

\item{max_r}{Numeric,}

\item{cellsize}{Numeric}

\item{base}{Raster,}
}
\value{
Raster
}
\description{
Create ConNest Probability Raster
}
\details{
CreateConNestProb(con_nest_raster, gamma_shape, gamma_rate, x, y, max_r,
  cellsize, base)
}
