% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home.R
\name{CreateHomeRangeKernels}
\alias{CreateHomeRangeKernels}
\title{CreateHomeRangeKernels}
\usage{
CreateHomeRangeKernels(df_all, df_home, base, max_r, home_inflection,
home_scale, avoid_inflection, avoid_scale, output_dir, write_distance,
write_homerange)
}
\arguments{
\item{df_all}{dataframe of all homerange centroids}

\item{df_home}{dataframe of homerange centroids to calculate homerange
kernels (these must be a subset of the df_all dataframe), Default is to use
'df_all' dataframe}

\item{base}{base Raster that sets the projection, extent, and dimensions of
the study area}

\item{max_r}{maximum radius to calculate the homerange raster from}

\item{home_inflection}{inflection point of the Logistic function that governs
the home kernel}

\item{home_scale}{scale parameter of the Logistic function that governs the
scale parameter}

\item{avoid_inflection}{inflection point of the Logistic function that
governs the conspecific avoidance kernel}

\item{avoid_scale}{scale parameter of the Logistic function that governs the
conspecific avoidance kernel}

\item{output_dir}{directory for output files (distance, homerange)}

\item{write_distance}{logical, write distance raster to file. Default is
FALSE.}

\item{write_homerange}{logical, write home range raster to file. Default is
FALSE.}
}
\value{
List, contains homerange kernel Rasters for all the df_home
  centroids
}
\description{
Creates RasterLayers of homerange kernels based on home centroids
}
