% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{CreateMoveKernelWeibull}
\alias{CreateMoveKernelWeibull}
\title{CreateMoveKernelWeibull}
\usage{
CreateMoveKernelWeibull(max_r, cellsize, mu, rho, shape, scale,
   ignore_cauchy, ignore_weibull)
}
\arguments{
\item{max_r}{maximum radius of kernel in meters, default = 300}

\item{cellsize}{cell size in meters, default = 30}

\item{mu}{mu parameter of wrapped Cauchy distribution, 0 radians is due east
because everything is based on the Unit Circle}

\item{rho}{rho parameter of wrapped Cauchy distribution}

\item{shape}{shape parameter of Pareto distribution}

\item{scale}{scale parameter of Pareto distribution}

\item{ignore_cauchy}{logical, removes cauchy kernel's contribution to output
raster. Default is FALSE.}

\item{ignore_weibull}{logical, removes Weibull kernel's contribution to
output raster. Default is FALSE.}
}
\value{
matrix
}
\description{
Create a movement kernel matrix based on a wrapped Cauchy distribution
  for direction and a Weibull distribution for distance.
}
\details{
the Weibull parameters need to be based on distance in meters
}
