% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{CreateMoveKernelWeibullVonMises}
\alias{CreateMoveKernelWeibullVonMises}
\title{CreateMoveKernelWeibullVonMises}
\usage{
CreateMoveKernelWeibullVonMises(max_r, cellsize, mu, rho, shape,
   scale, ignore_cauchy, ignore_weibull)
}
\arguments{
\item{max_r}{maximum radius of kernel in meters, default = 300}

\item{cellsize}{cell size in meters, default = 30}

\item{mu1}{mu1 parameter of mixed von Mises distribution, 0 radians is due
East because everything is based on the Unit Circle}

\item{mu2}{mu2 parameter of mixed von Mises distribution, 0 radians is due
East because everything is based on the Unit Circle}

\item{kappa1}{kappa1 parameter of mixed von Mises distribution}

\item{kappa2}{kappa2 parameter of mixed von Mises distribution}

\item{mix}{mixture (p) parameter of mixed von Mises distribution}

\item{shape}{shape parameter of Weibull distribution}

\item{scale}{scale parameter of Weibull distribution}

\item{ignore_von_mises}{logical, removes mixed von Mises kernel's
contribution to output raster. Default is FALSE.}

\item{ignore_weibull}{logical, removes Weibull kernel's contribution to
output raster. Default is FALSE.}
}
\value{
matrix
}
\description{
Create a movement kernel matrix based on a mixed von Mises distribution
  for direction and a Weibull distribution for distance.
}
\details{
the Weibull parameters need to be based on distance in meters
}
