% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{CreateParetoKernel}
\alias{CreateParetoKernel}
\title{CreateParetoKernel}
\usage{
CreateParetoKernel(scale, shape, max_r, cellsize)
}
\arguments{
\item{scale}{scale parameter of Pareto distribution}

\item{shape}{shape parameter of Pareto distribution}

\item{max_r}{maximum radius of kernel in meters, default = 100}

\item{cellsize}{cell size in meters, default = 1}
}
\value{
RasterLayer
}
\description{
Creates a probability matrix based on a Pareto distribution.
}
