% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{CreateParsClassConstant}
\alias{CreateParsClassConstant}
\title{CreateParsClassConstant}
\usage{
CreateBirthDate(step_cauchy_mu, step_cauchy_rho)
}
\arguments{
\item{step_cauchy_mu}{step cauchy mu}

\item{step_cauchy_rho}{step cauchy rho}
}
\value{
list
}
\description{
Creates a parameter class 'constant'
}
