% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{CreateParsClassSeason}
\alias{CreateParsClassSeason}
\title{CreateParsClassSeason}
\usage{
CreateParsClassSeason(step_pareto_scale,
  step_pareto_shape, step_max_r)
}
\arguments{
\item{step_pareto_scale}{pareto scale}

\item{step_pareto_shape}{pareto shape}

\item{step_max_r}{max radius}
}
\value{
list
}
\description{
Creates a parameter class 'season'
}
