% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{CreateParsGlobal}
\alias{CreateParsGlobal}
\title{CreateParsClassSeason}
\usage{
CreateParsClassSeason(input)
}
\arguments{
\item{sim_start}{POSIXct default = as.POSIXct("2017-01-01", tz = "UTC")}

\item{sim_period}{'period', default = period(100, "days")}

\item{sim_end}{POSIXct, default = as.POSIXct("2017-01-01", tz = "UTC")}

\item{rep_period}{'period', default = period(1, "days")}

\item{rep_interval}{character string of dates, default = c("01Jan", "01Jul")}

\item{rep_interval_custom}{= character string of dates, dafault = NULL}

\item{step_period}{'period', default = period(1, "day")}

\item{time_step_period}{'period', default = period(6, "hour")}

\item{birth_day}{= character string of date, default = "01Jan"}

\item{input_age_period}{= character string of period, default = "years"}

\item{report_age_period}{= character string of period, default = "months"}

\item{sim_seasons}{= dataframe, columns = c(start, season), default are
typical sesasons}
}
\value{
list
}
\description{
Creates a parameter class 'season'
}
