% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{CreateReportIntervals}
\alias{CreateReportIntervals}
\title{CreateReportIntervals}
\usage{
CreateReportIntervals(sim)
}
\arguments{
\item{sim}{a list that contains a "pars" list which contains a "global" list
that sets the arguments for the Report Interval. See Description section
below.}
}
\value{
list of intervals
}
\description{
sim list that contains a "pars" list which contains a "global" list that
contains the following objects:
  sim_start = a POSIXct object for simulation's start, required
  sim_period = a Period object (from 'lubridate' package), e.g. period(10,
    "year"), period(10, "week"), period(10, "week") that sets the length of
    time the simulation runs
  sim_end = a POSIXct object for simulation's end. Setting this will override
    the 'sim_period' parameter. Default = NULL.
  rep_period = a Period object that sets the summary interval
  rep_interval = vector or dateframe (with date in first column) of the
    format %B%d, %b%d, %d%B, or %B%d (see ?strptime for more details on
    formatting). Example: c("April01", "May15", "Sep1", "Oct15"). Overrides
    'rep_period' parameter.
  rep_interval_custom = a vector of POSIXct objects that set the start of the
    summary periods. If sim_start precedes the first value, the first value
    will go until from start_sim to rep_interval_custom. Last interval will
    go from the last date in rep_interval_custom to end of sim_period or
    end_sim. Overrides 'rep_period' or 'rep_interval' parameters.
}
\details{
Creates summary intervals for the simulation input
}
\note{
Either 'sim_period' or 'sim_end' must be set, and either 'rep_period',
'rep_interval', or 'rep_interval_custom' must be set. For 'rep_interval' the
dates can be set as numeric for both day and month, but there is a
possibility that the day and month will be inverted because dates with both
values <12 can be confused (e.g. 02-10 can be either February 10th or October
2nd. Date format: "10Feb" or "02Oct" is preferred.
}
