% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{CreateRunsList}
\alias{CreateRunsList}
\title{CreateRunsList}
\usage{
CreateRunsList(runs = runs)
}
\arguments{
\item{runs}{= numeric, number of runs in RunSimulation()}
}
\value{
a list
}
\description{
Helper function for RunSimulation(). Creates a list of empty objects, each
one named individually as "run_xx" where the number of leading zeroes matches
the maximum digit width of the number of runs.
}
\note{
Where there are < 10 runs, there will have no leading zeros, when
there are between 10 and 99 runs, there will be a leading zero for runs
numbered < 10.
}
