% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{CreateSimConNestDistRasters}
\alias{CreateSimConNestDistRasters}
\title{Create con_nest distance rasters for simulation}
\usage{
CreateSimConNestDistRasters(
  agents,
  nest_set,
  base = base,
  output_dir = "Output/Sim/Territorial",
  max_r = 30000,
  write_con_nest_all = TRUE
)
}
\arguments{
\item{agents}{list, agents}

\item{nest_set}{dataframe, nests with coordinates}

\item{base}{raster, the base raster layer for the analysis}

\item{output_dir}{character, location for the output}

\item{max_r}{numeric, maximum radius for kernels. Default = 30000}

\item{write_con_nest_all}{logical, whether or not to write the}
}
\description{
Create con_nest distance rasters for simulation
}
