% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{CreateStepIntervals}
\alias{CreateStepIntervals}
\title{CreateStepIntervals}
\usage{
CreateStepIntervals(rep_interval, step_period)
}
\arguments{
\item{rep_interval}{= an Interval object from lubridate that represents the
current summary interval}

\item{step_period}{= a Period object specifying the length of the step
period. The step_period should be of equal or greater length than the
time_step_period.}
}
\value{
a list of intervals
}
\description{
Create step intervals within a summary interval
}
