% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{CreateTimeSteps}
\alias{CreateTimeSteps}
\title{CreateTimeSteps}
\usage{
CreateTimeSteps(step_interval)
}
\arguments{
\item{step_interval}{= an Interval object from lubridate that represents the
current step interval}

\item{time_step_period}{= a Period object specifying the length of the time
step period. Note that if the length of the current_step_interval is equal to
the time_step_period the function will return a POSIXct object with the same
date as the end of the current_step_interval.}
}
\value{
a list of intervals
}
\description{
Create the time steps within a step interval
}
