% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{CreateTimeStepsBAEA}
\alias{CreateTimeStepsBAEA}
\title{CreateTimeStepsBAEA}
\usage{
CreateTimeStepsBAEA(step_interval)
}
\arguments{
\item{step_interval}{= an Interval object from lubridate that represents the
current step interval}

\item{sim}{= 'sim' list}
}
\value{
a list of intervals
}
\description{
Create the time steps within a step interval based on sunrise and sunset.
Specifically designed for BAEA project.
}
