% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{CreateTimeStepsInStepIntervalBAEA}
\alias{CreateTimeStepsInStepIntervalBAEA}
\title{CreateTimeStepsinStepIntervalBAEA}
\usage{
CreateTimeStepsinStepIntervalBAEA(step_interval, sim)
}
\arguments{
\item{step_interval}{= an Interval object from lubridate that represents the
current step interval}

\item{sim}{= 'sim' list}
}
\value{
a list of intervals
}
\description{
Create the time steps within a step interval based on sunrise and sunset in
  Bangor, ME. Specifically designed for BAEA project.
}
