% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{CustomAgentsReportData}
\alias{CustomAgentsReportData}
\title{CustomAgentsReportData}
\usage{
CustomAgentsReportData(agent_states, report_data, step_data, q)
}
\arguments{
\item{agent_states}{= agent_states}

\item{report_data}{= report_data}

\item{step_data}{= step_data}

\item{q}{= counter for reporting interval}
}
\value{
'report_data' object
}
\description{
Creates custom report data
}
