% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{CustomPopReportData}
\alias{CustomPopReportData}
\title{CustomPopReportData}
\usage{
CustomPopReportData(report_data, q, agent_states, step_data)
}
\arguments{
\item{pop_report}{= population report}

\item{report_data}{=  report_data}

\item{q}{= report_data row for new data}
}
\value{
a population report
}
\description{
Creates and updates custom pop_report data.
}
