% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{ExtractUnitFromPeriod}
\alias{ExtractUnitFromPeriod}
\title{ExtractUnitFromPeriod}
\usage{
ExtractUnitFromPeriod(period)
}
\arguments{
\item{period}{= period object}
}
\value{
period object's unit as a chararcter
}
\description{
Helper function for CreateSumInterval(). Finds unit of Period object.
}
