% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{FindFirstReportInterval}
\alias{FindFirstReportInterval}
\title{FindFirstReportInterval}
\usage{
FindFirstReportInterval(sim_start, rep_interval)
}
\arguments{
\item{sim_start}{= a POSIXct object for simulation's start, required}

\item{rep_interval}{= vector or dateframe (with date in first column) of the
format %B%d, %b%d, %d%B, or %B%d (see ?strptime for more details on
formatting). Example: c("April01", "May15", "Sep1", "Oct15"). Required.}
}
\value{
an Interval object
}
\description{
Helper function for CreateSumInterval. Finds first interval based on
sim_start datetime and rep_interval dates
}
