% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{FindSeasonFromDatetime}
\alias{FindSeasonFromDatetime}
\title{FindSeasonFromDatetime}
\usage{
FindSeasonFromDatetime(datetime, season)
}
\arguments{
\item{datetime}{= a POSIXct object}

\item{seasons}{= a dataframe of starting dates ("start" column) and season
names ("season" column), usually located at: sim$pars$global$sim_seasons}
}
\value{
an atomic character object
}
\description{
Determines which season (from sim$pars$global$sim_seasons) a datetime is
within
}
