% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibmr.R
\docType{package}
\name{ibmr}
\alias{ibmr}
\title{ibmr: An individual-based model (IBM) framework for conducting
  spatially-explicit simulation.}
\description{
This package has the basic structure to run an IBM, but will need to modified
by the user in order to make a useful model. Still in DEVELOPMENT phase.
}
\section{List of 'sim' Components}{

The 'sim' list is the fundamental components of the IBM simulation.
}

