% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{LastDailySubModelBAEA}
\alias{LastDailySubModelBAEA}
\title{LastDailySubModelBAEA}
\usage{
LastDailySubModelBAEA(sim, agent_states, step_data, step)
}
\arguments{
\item{sim}{sim object}

\item{agent_states}{agent_states object}

\item{step_data}{step_data object}

\item{step}{step object}
}
\value{
step_data
}
\description{
Ensures that the last location in a day has a behavior of either "nest" or
"roost".
}
