% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{MovementSubModelBAEA}
\alias{MovementSubModelBAEA}
\title{MovementSubModelBAEA}
\usage{
MovementSubModelBAEA(sim, agent_states, step_data, step)
}
\arguments{
\item{sim}{'sim' object}

\item{agent_states}{'agent_states' list object}

\item{step_data}{'step_data' dataframe}

\item{step}{step interval}
}
\value{
'step_data' object
}
\description{
Movement submodel that predicts movements based on current location,
  homerange_kernel, nest_return probability, and others.
}
