% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{NamedList}
\alias{NamedList}
\title{NamedList}
\usage{
NamedList(...)
}
\arguments{
\item{...}{= objects to add to list}
}
\value{
a list
}
\description{
Creates a list with named objects, and it tries not to replace any
already-named arguments.
}
\note{
Original function came from Ben Bolker's answer on StackOverflow:
"http://stackoverflow.com/questions/16951080"
}
