% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{NonlinearRangeRescaleGamma}
\alias{NonlinearRangeRescaleGamma}
\title{NonlinearRangeRescaleGamma}
\usage{
NonlinearRangeRescaleGamma(x, shape = shape, rate = rate, min = NULL,
  max = NULL, lowBound = 1, upBound = NULL,
  movement_kernel = movement_kernel, negative = TRUE)
}
\arguments{
\item{x}{numeric}

\item{shape}{numeric}

\item{rate}{numeric}

\item{min}{numeric}

\item{max}{numeric}

\item{lowBound}{numeric}

\item{upBound}{numeric}

\item{movement_kernel}{kernel}

\item{negative}{logical}
}
\value{
vector
}
\description{
Non-linear Range Rescale Gamma Distribution
}
