% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{NonlinearRangeRescalePareto}
\alias{NonlinearRangeRescalePareto}
\title{NonlinearRangeRescalePareto}
\usage{
NonlinearRangeRescalePareto(x, scale = shape, shape = shape,
  min_prob = NULL, max_prob = NULL, lower_bound = 1,
  upper_bound = NULL, movement_kernel = movement_kernel,
  negative = TRUE)
}
\arguments{
\item{x}{numeric, x (independent variable) value}

\item{scale}{numeric, scale (sigma) parameter for texmex::gpd}

\item{shape}{numeric, shape (xi) parameter for texmex::gpd}

\item{min_prob}{numeric, minimum Pareto probability value}

\item{max_prob}{numeric, maximum Pareto probability value}

\item{lower_bound}{numeric, lower limit for rescaled y values}

\item{upper_bound}{numeric, upper limit for rescaled y values}

\item{movement_kernel}{kernel, may be used to get the upper_bound}

\item{negative}{logical, whether or not to multiple outcome by -1}
}
\value{
vector
}
\description{
Non-linear Range Rescale Pareto Distribution
}
