% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home.R
\name{PlotHomeRange}
\alias{PlotHomeRange}
\title{PlotHomeRange}
\usage{
PlotHomeRange(homerange, col, extent)
}
\arguments{
\item{homerange}{home range probability raster}

\item{col}{color palette for base raster, default is jet2.col()}

\item{extent}{an extent object for the plot, default is homerange extent}
}
\value{
plot of locations and pathways
}
\description{
Plot a homerange raster
}
