% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home.R
\name{PlotSimHomeRange}
\alias{PlotSimHomeRange}
\title{PlotSimHomeRange}
\usage{
PlotSimHomeRange(sim, home_df, homerange, col, label, extent)
}
\arguments{
\item{sim}{dataframe with x, y columns}

\item{df_home}{dataframe with x, y columns}

\item{homerange}{home range probability raster}

\item{col}{color palette for base raster, default is jet2.col()}

\item{label}{column name for point labels}

\item{extent}{an extent object for the plot, default is homerange extent}
}
\value{
plot of locations and pathways
}
\description{
Plot locations, pathways, and bias values of SimulateHomeRange() output
dataframes
}
