% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBPRESSO.R
\name{print.mrpresso_ib}
\alias{print.mrpresso_ib}
\title{Print method for mrpresso_ib}
\usage{
\method{print}{mrpresso_ib}(x, ...)
}
\arguments{
\item{x}{An object of class mrpresso_ib}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for mrpresso_ib
}
