% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{RemoveExcept}
\alias{RemoveExcept}
\title{RemoveExcept}
\usage{
RemoveExcept(object)
}
\arguments{
\item{object}{= objects to keep}
}
\value{
environment with only the kept objects
}
\description{
Helper function that removes all the objects in the environment except ones
listed in argument
}
