% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{ReproductionSubModel}
\alias{ReproductionSubModel}
\title{ReproductionSubModel}
\usage{
ReproductionSubModel(agent_states = agent_states,
  step_data = step_data)
}
\arguments{
\item{agent_states}{= a 'agent_states' list object}

\item{step_data}{= list object}
}
\value{
'agent_states' list object
}
\description{
Reproduction submodel
}
\note{
Just a placeholder for now.
}
