% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{ReturnActiveSeq}
\alias{ReturnActiveSeq}
\title{ReturnActiveSeq}
\usage{
ReturnActiveSeq(all = sim$agents$all, rep_interval = rep_interval)
}
\arguments{
\item{all}{= 'all' list}

\item{rep_interval}{= reporting interval}
}
\value{
'active_seq' vector
}
\description{
Create vector of positions of agents that had a step_data$datetime within
the current reporting interval
}
