% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{ReturnAliveIds}
\alias{ReturnAliveIds}
\title{ReturnAliveIds}
\usage{
ReturnAliveIds(all = sim$agents$all)
}
\arguments{
\item{all}{= 'all' list}
}
\value{
a vector
}
\description{
Create vector of ids of alive agents
}
