% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{ReturnAliveSeq}
\alias{ReturnAliveSeq}
\title{ReturnAliveSeq}
\usage{
ReturnAliveSeq(sim = sim)
}
\arguments{
\item{sim}{= 'sim' list}
}
\value{
a vector
}
\description{
Create vector of alive agents positions
}
