% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{ReturnPopReportData}
\alias{ReturnPopReportData}
\title{ReturnPopReportData}
\usage{
ReturnPopReportData(sim, agents, rep_interval)
}
\arguments{
\item{sim}{= 'sim' list}

\item{agents}{= agents object}

\item{rep_interval}{= counter for reporting interval (q)}
}
\value{
period object's unit as a chararcter
}
\description{
Helper function for UpdatePopReport(). Finds intervals with start dates
within the report interval.
}
