% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{ReturnReportStepData}
\alias{ReturnReportStepData}
\title{ReturnReportStepData}
\usage{
ReturnReportStepData(agent = agent, rep_interval = rep_interval)
}
\arguments{
\item{agent}{agent = 'agent' list}

\item{rep_interval}{= counter for reporting interval (q)}
}
\value{
a step dataframe
}
\description{
Returns agent's step data within the current reporting interval
}
