% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{RunSimulation}
\alias{RunSimulation}
\title{RunSimulation}
\usage{
RunSimulation(sim, runs, write, output_dir)
}
\arguments{
\item{sim}{= list of (agents, pars, and spatial)}

\item{runs}{= number of runs}

\item{write}{= write}

\item{output_dir}{= output files directory}
}
\value{
A list object
}
\description{
The backbone function for an individual-based model simulation.
}
