% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home.R
\name{SampleRandomPointsInHomerange}
\alias{SampleRandomPointsInHomerange}
\title{SampleRandomPointsInHomerange}
\usage{
SampleRandomPointsInHomerange (df_all, df_home, used_pts, base,
  max_r, id, name, output_dir)
}
\arguments{
\item{df_all}{dataframe of all homerange centroids}

\item{df_home}{dataframe of homerange centroids to calculate homerange
kernels (these must be a subset of the df_all dataframe), Default is to
use 'df_all' dataframe}

\item{used_pts}{dataframe of points to use for analysis}

\item{base}{Raster that sets the projection, extent, and dimensions of the
study area}

\item{max_r}{maximum radius to calculate the homerange raster from each
df_home centroid}

\item{id}{column name of df_home that identifies the homerange. Default is
NULL, which sets the names to df_home row number.}

\item{name}{column name of df_home that identifies the name of the homerange
and is used to write the file name of the .tif.Default is 'id', which sets
the names to df_home row number if 'id' is NULL.}

\item{output_dir}{directory for output files (, homerange)}

\item{write_homerange}{logical, write home range raster to file. Default is
FALSE.}
}
\value{
A list containing homerange kernel Rasters for all the df_home
  centroids.
}
\description{
Creates RasterLayers of homerange kernels based on home centroids
}
\details{
A list containing homerange kernel Rasters for all the df_home
  centroids.
}
