% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{SurvivalSubModel}
\alias{SurvivalSubModel}
\title{SurvivalSubModel}
\usage{
SurvivalSubModel(agent_states = agent_states, step_data = step_data)
}
\arguments{
\item{agent_states}{= 'agent_states' list object}

\item{step_data}{= list object}
}
\value{
agent_states
}
\description{
Survival submodel
}
\note{
Just a placeholder for now.
}
