% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{UpdateAgentParsData}
\alias{UpdateAgentParsData}
\title{UpdateAgentParsData}
\usage{
CreateAgentParsData(sim, init)
}
\arguments{
\item{sim}{= sim list, MUST be included in function's parameter argument
needed when: when init == TRUE}

\item{init}{= logical, whether or not this is the initation step}
}
\value{
'sim' list
}
\description{
Creates and updates agents parameters dataframe
}
