% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{UpdateAgentsReport}
\alias{UpdateAgentsReport}
\title{UpdateAgentsReport}
\usage{
CreateAgentInData(agents, init)
}
\arguments{
\item{sim}{= 'sim' list}

\item{rep_interval}{= rep_interval}

\item{step_intervals}{= step_intervals}
}
\value{
an 'agents' list
}
\description{
Creates and updates agents interval data dataframe
}
\note{
rep_data, init are internal to RunSimualtion()
}
