% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{UpdateAgentStates}
\alias{UpdateAgentStates}
\title{UpdateAgentStates}
\usage{
UpdateAgentStates(agent_states = NULL, sim = sim, init = FALSE)
}
\arguments{
\item{agent_states}{a list from sim$agents$all$agent$states}

\item{sim}{sim list, MUST be included in function's parameter arguments
needed when: when init == TRUE}

\item{init}{logical, whether or not this is the initation step}
}
\value{
an 'agents' list
}
\description{
Creates and updates agents list from the sim object
}
