% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{UpdateAgentStepData}
\alias{UpdateAgentStepData}
\title{UpdateAgentStepData}
\usage{
CreateAgentStepData(agents, sim, init)
}
\arguments{
\item{step_data}{= 'step data' dataframe}

\item{sim}{= 'sim' object}

\item{init}{= logical, whether or not this is the initation step}
}
\value{
an 'agents' list
}
\description{
Creates and updates agents step data database
}
