% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{UpdateAgentStepDataBAEA}
\alias{UpdateAgentStepDataBAEA}
\title{UpdateAgentStepDataBAEA}
\usage{
UpdateAgentStepDataBAEA(step_data, sim, time_step_dfinit)
}
\arguments{
\item{step_data}{= 'step data' dataframe}

\item{sim}{= 'sim' object}

\item{init}{= logical, whether or not this is the initation step}

\item{rep_intervals}{= 'rep_intervals' object}
}
\value{
an 'step_data' dataframe
}
\description{
Creates and updates agents step data dataframes with timesteps based on
starting locations of each agent and dates for the sim so that the timesteps
are similar to the duty cycle of the recorded data: 15-minute time steps
starting an hour before sunrise and going until an hour after sunset.
}
