% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{UpdatePopReport}
\alias{UpdatePopReport}
\title{UpdatePopReport}
\usage{
UpdatePopReportData(sim, rep_interval, step_intervals)
}
\arguments{
\item{sim}{'sim' list}

\item{rep_interval}{rep_interval}

\item{step_intervals}{step_intervals}
}
\value{
a 'pop_report' dataframe
}
\description{
Creates and updates population interval dataframe
}
