% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{WriteSimList}
\alias{WriteSimList}
\title{WriteSimList}
\usage{
WriteSimList(write, run, sim, output_dir, components)
}
\arguments{
\item{write}{= logical, whether or not to write the sim list to a file.
Default is TRUE.}

\item{run}{= name of run. Default uses the name from the runs object to
ensure that the proper number of leading zeros is used. If default is not
use, the name will simply be: "sim_(run).RData"}

\item{sim}{= a 'sim' list object}

\item{output_dir}{= output directory, default is working directory}

\item{components}{= components of 'sim' to write. Options include: "agents",
"pars", and "spatial". Default is all.}
}
\value{
Writes a file to the output_dir
}
\description{
Writes "sim" list to a directory,
}
\note{
NOT FINISHED - Code not implemented for writing subset of components
}
