% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{CenterXYInCell}
\alias{CenterXYInCell}
\title{CenterXYInCell}
\usage{
CenterXYInCell(x, y, xmin, ymin, cellsize)
}
\arguments{
\item{x}{x value}

\item{y}{y value}

\item{xmin}{minimum value of x that establishes the grid arrangement}

\item{ymin}{minimum value of y that establishes the grid arrangement}

\item{cellsize}{cell size in units of the x and y values}

\item{return}{character, what to return ("x", "y", or "xy"). "xy" is default}
}
\value{
vector x, y, or c(x,y)
}
\description{
Centers x and y values into the center of a raster cell based on the xmin,
ymin, and cellsize parameters
}
